Code restructure and styling updates
The following is current progress:
Small steps to start using Moodle templates.

Added work around to fix some styling issues that occurred when using Moodle build in Form API, the work around is aiming to solve the Label not probably displayed based on the user’s screen size.

Started the proceed to do some code refactoring to help with future updates to improve on it or help with Unit Testing.

updated version file to reflect current release build and plugin
maturity level